code for booking in separate modules
